requirements.txt
google-generativeai>=0.3.0
Pillow>=9.0.0
IPython>=7.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
kokoro>=0.9.2
soundfile>=0.10.0
numpy>=1.20.0
requests>=2.0.0
